feat(deploy/ha): Add stubs for getting yamls from ./halconfig.
